This change defines DOUBLE as the wire format for floats, since we have no way to truly parse or produce a float on the client side.  Fixes failing RPCSuite web.
Review by: mmendez


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2194 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/server/rpc/impl/ServerSerializationStreamReader.java
@@ -329,7 +329,7 @@ public final class ServerSerializationStreamReader extends
   }
 
   public float readFloat() {
-    return Float.parseFloat(extract());
+    return (float) Double.parseDouble(extract());
   }
 
   public int readInt() {
